splitting the functions in the officer_management.py into their own separate files
